Rename artifact.rs into oci_artifact.rs
