+++ ocipkg/src/image/mod.rs
@@ -4,18 +4,18 @@
 
 pub mod annotations;
 
-mod artifact;
 mod config;
 mod layout;
 mod oci_archive;
+mod oci_artifact;
 mod oci_dir;
 mod read;
 mod write;
 
-pub use artifact::*;
 pub use config::*;
 pub use layout::*;
 pub use oci_archive::*;
+pub use oci_artifact::*;
 pub use oci_dir::*;
 pub use read::*;
 pub use write::*;
